move reductions to separate file
